backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
sqlalchemy==2.0.36
alembic==1.13.3
httpx>=0.27
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tenacity>=8.2
python-dotenv==1.0.1
psycopg2-binary==2.9.10
PyJWT>=2.8
aiofiles>=23.2.1

# Optional for YouTube
google-api-python-client>=2.131
google-auth>=2.33
google-auth-oauthlib>=1.2

# Optional scheduling
apscheduler>=3.10